refactor(navbar-comp): move comp and css file into own folder
