+++ src/pages/homepage/index.tsx
@@ -3,7 +3,7 @@ import { useAuth } from '../../auth/useAuth';
 import { UserEmail } from '../../types';
 import Header from '../../components/header';
 import MovieList from './MovieList';
-import Navbar from '../../components/Navbar';
+import Navbar from '../../components/navbar/Navbar';
 
 const Homepage = () => {
   return (
