+++ src/pages/savedMovies/index.tsx
@@ -10,7 +10,7 @@ import { useAuth } from '../../auth/useAuth';
 import { getImgUrl } from '../../api';
 import { useFullMovies } from '../homepage/hooks';
 import DropdownMenu from '../homepage/DropdownMenu';
-import Navbar from '../../components/Navbar';
+import Navbar from '../../components/navbar/Navbar';
 import styles from '../homepage/MovieList.module.css';
 
 const SavedMovies = () => {
